yaomefilm/requirements.txt
django
django-filter
psycopg2-binary
